chore(scp): move tar operations to internal/filecopy/
