+++ internal/filecopy/doc.go
@@ -1,0 +1,6 @@
+// Copyright 2026 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+// Package filecopy contains the tar archive handling used by
+// the Kubernetes provider and Juju client for SCP to K8s pods.
+package filecopy
